+++ src/tui/add.rs
@@ -7,11 +7,11 @@ use safe_path::scoped_join;
 use std::path::PathBuf;
 
 use crate::actions::{
-  Action, DescribedAction, buildlike::*, check::CheckAction, deploylike::*, new_action, observe::ObserveAction,
-  packlike::*, patch::PatchAction, storage_add::AddToStorageAction,
+  Action, DescribedAction, buildlike::*, deploylike::*, new_action, observe::ObserveAction, packlike::*,
+  patch::PatchAction, storage_add::AddToStorageAction, test::TestAction,
 };
 use crate::cmd::{NewActionArgs, NewPipelineArgs};
-use crate::configs::{DeployerGlobalConfig, DeployerProjectOptions};
+use crate::configs::{DeployerGlobalConfig, DeployerProjectOptions, Placement};
 use crate::entities::{
   auto_version::AutoVersionExtractFromRule,
   custom_command::CustomCommand,
@@ -90,7 +90,6 @@ impl DescribedAction {
 
     let action_types: Vec<&str> = vec![
       "Custom",
-      "Check",
       "Use content from storage",
       "Patch",
       "Sync build folder to remote",
@@ -117,8 +116,8 @@ impl DescribedAction {
         let command = CustomCommand::new_from_prompt()?;
         Action::Custom(command)
       }
-      "Check" => Action::Check(CheckAction::new_from_prompt()?),
-      action_type @ ("Pre-build" | "Build" | "Post-build" | "Test") => {
+      "Test" => Action::Test(TestAction::new_from_prompt()?),
+      action_type @ ("Pre-build" | "Build" | "Post-build") => {
         let supported_langs = specify_programming_languages()?;
         let commands = collect_multiple_commands()?;
 
@@ -131,7 +130,6 @@ impl DescribedAction {
           "Pre-build" => Action::PreBuild(action),
           "Build" => Action::Build(action),
           "Post-build" => Action::PostBuild(action),
-          "Test" => Action::Test(action),
           _ => unreachable!(),
         }
       }
@@ -260,7 +258,7 @@ pub fn collect_requirements() -> anyhow::Result<Option<Vec<Requirement>>> {
   Ok(if reqs.is_empty() { None } else { Some(reqs) })
 }
 
-impl CheckAction {
+impl TestAction {
   pub fn new_from_prompt() -> anyhow::Result<Self> {
     let bash_c = specify_bash_c(None)?;
 
@@ -613,7 +611,7 @@ pub fn collect_variables() -> anyhow::Result<Vec<Variable>> {
   Ok(v)
 }
 
-pub fn collect_af_inplacement(artifacts: &[impl AsRef<str>]) -> anyhow::Result<(PathBuf, PathBuf)> {
+pub fn collect_af_inplacement(artifacts: &[impl AsRef<str>]) -> anyhow::Result<Placement> {
   use inquire::{Select, Text};
 
   let assume_root = PathBuf::from("/");
@@ -623,14 +621,14 @@ pub fn collect_af_inplacement(artifacts: &[impl AsRef<str>]) -> anyhow::Result<(
     let from = PathBuf::from(Select::new(i18n::SELECT_PROJECT_AF, artifacts.to_owned()).prompt()?);
     let to = PathBuf::from(Text::new(i18n::CHOOSE_AF_INPLACEMENT).prompt()?);
     if scoped_join(&assume_root, &to).is_ok() {
-      return Ok((from, to));
+      return Ok(Placement { from, to });
     } else {
       println!("{}", i18n::INCORRECT_AF_INPL_PATH)
     }
   }
 }
 
-pub fn collect_af_inplacements(artifacts: &[PathBuf]) -> anyhow::Result<Vec<(PathBuf, PathBuf)>> {
+pub fn collect_af_inplacements(artifacts: &[PathBuf]) -> anyhow::Result<Vec<Placement>> {
   use inquire::Confirm;
 
   let artifacts = artifacts.iter().map(|v| v.to_str().unwrap()).collect::<Vec<_>>();
@@ -727,11 +725,15 @@ impl Variable {
   }
 
   pub fn new_plain_from_prompt() -> anyhow::Result<VarValue> {
-    Ok(VarValue::Plain(inquire::Text::new(i18n::VAR_PLAIN_CONTENT).prompt()?))
+    Ok(VarValue::Plain {
+      value: inquire::Text::new(i18n::VAR_PLAIN_CONTENT).prompt()?,
+    })
   }
 
   pub fn new_env_from_prompt() -> anyhow::Result<VarValue> {
-    Ok(VarValue::FromEnvVar(inquire::Text::new(i18n::VAR_ENV_KEY).prompt()?))
+    Ok(VarValue::FromEnvVar {
+      var_name: inquire::Text::new(i18n::VAR_ENV_KEY).prompt()?,
+    })
   }
 
   pub fn new_env_file_from_prompt() -> anyhow::Result<VarValue> {
@@ -743,7 +745,7 @@ impl Variable {
 
   pub fn new_kv2_from_prompt() -> anyhow::Result<VarValue> {
     println!("{}: {}", i18n::NOTE.green().italic(), i18n::KV2_NOTE);
-    Ok(VarValue::FromHCVaultKv2(Kv2Paths {
+    Ok(VarValue::FromHcVaultKv2(Kv2Paths {
       mount_path: inquire::Text::new(i18n::VAR_MOUNT_PATH).prompt()?,
       secret_path: inquire::Text::new(i18n::VAR_SECRET_PATH).prompt()?,
     }))
@@ -769,21 +771,25 @@ impl Requirement {
   }
 
   pub fn new_exists_from_prompt() -> anyhow::Result<Self> {
-    Ok(Self::Exists(collect_path()?))
+    Ok(Self::Exists { path: collect_path()? })
   }
 
   pub fn new_exists_any_from_prompt() -> anyhow::Result<Self> {
-    Ok(Self::ExistsAny(collect_paths()?))
+    Ok(Self::ExistsAny {
+      paths: collect_paths()?,
+    })
   }
 
   pub fn new_check_from_prompt() -> anyhow::Result<Self> {
-    Ok(Self::CheckSuccess(CheckAction::new_wop_from_prompt()?))
+    Ok(Self::CheckSuccess {
+      action: TestAction::new_wop_from_prompt()?,
+    })
   }
 
   pub fn new_remote_from_prompt() -> anyhow::Result<Self> {
-    Ok(Self::RemoteAccessibleAndReady(ShortName::new(
-      inquire::Text::new(i18n::REMOTE_SHORT_NAME).prompt()?,
-    )?))
+    Ok(Self::RemoteAccessibleAndReady {
+      remote_host_name: ShortName::new(inquire::Text::new(i18n::REMOTE_SHORT_NAME).prompt()?)?,
+    })
   }
 }
 
@@ -803,10 +809,7 @@ impl TargetDescription {
       "Android" => OsVariant::Android,
       "iOS" => OsVariant::iOS,
       "Linux" => OsVariant::Linux,
-      "Unix-like" => {
-        let name = Text::new(i18n::TARGET_OS_UNIX_LIKE).prompt()?;
-        OsVariant::UnixLike(name)
-      }
+      "Unix-like" => OsVariant::UnixLike,
       "Windows" => OsVariant::Windows,
       "macOS" => OsVariant::macOS,
       "Other" => {
@@ -816,7 +819,7 @@ impl TargetDescription {
       _ => unreachable!(),
     };
 
-    let derivative = Text::new(i18n::TARGET_OS_DER).prompt()?;
+    let os_derivative = Text::new(i18n::TARGET_OS_DER).prompt()?;
 
     let version_type = Select::new(
       i18n::TARGET_OS_VER_S,
@@ -824,15 +827,15 @@ impl TargetDescription {
     )
     .prompt()?;
 
-    let version = match version_type {
+    let os_version = match version_type {
       i18n::TARGET_OS_VER_NS => OsVersionSpecification::No,
       i18n::TARGET_OS_VER_WS => {
         let ver = Text::new(i18n::TARGET_OS_VER).prompt()?;
-        OsVersionSpecification::Weak(ver)
+        OsVersionSpecification::Weak { version: ver }
       }
       i18n::TARGET_OS_VER_SS => {
         let ver = Text::new(i18n::TARGET_OS_VER).prompt()?;
-        OsVersionSpecification::Strong(ver)
+        OsVersionSpecification::Strong { version: ver }
       }
       _ => unreachable!(),
     };
@@ -840,8 +843,8 @@ impl TargetDescription {
     Ok(TargetDescription {
       arch,
       os: os_variant,
-      derivative,
-      version,
+      os_derivative,
+      os_version,
     })
   }
 }